requirements.txt
# Core dependencies
streamlit>=1.31.0
extra-streamlit-components>=0.1.60
chromadb>=0.4.22
anthropic>=0.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
watchdog>=3.0.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pre-commit>=3.6.0
ruff>=0.8.0
pydocstyle>=6.3.0
bandit>=1.7.6
